auth-service/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
flask-cors==4.0.0
